+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 2.5
+* Essentials.Pathfinding: NavMeshPath DataObject changed to ListCopyable
 * GOAPAction: fixed point math support added [Oleg-Grim]
 * GetNearest: distance dependant flag added [Oleg-Grim]
 * OverrideCost: implemented [Oleg-Grim]